sigma factor not carried through to derivatives
There is a factor of a half from the pair of equivalent lines in Hugenholtz self-energy diagram in the sigma terms. E occurs in the denominator so derivative with respect to E will be of form -E^-2, so the overall factor will be - half. Defined constant definition for au to eV in one place rather than scattered throughout code.